+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [3.2.0] - 2024-12-28
+### Added
+- New [toString](./src//Terms/DateTimeInterval.php) method, helps easier make PHP date time string
+- 
+### Fixed
+- Minor differential errors
+
 ## [3.1.2] - 2024-12-25
 ### Fixed
 - Known issues [Error: calculating next Solar term #87](https://github.com/luc-nham/lunar-calendar/issues/89)